+++ src/jvmMain/kotlin/Class340.kt
@@ -1,13 +1,12 @@
 import Class348_Sub1_Sub2.Companion.method2730
-import java.math.BigInteger
 
-class Class340 internal constructor(class248: Class248?, class112: Class112?, biginteger: BigInteger?, biginteger_14_: BigInteger?) {
+class Class340 internal constructor(class248: Class248?, class112: Class112?, biginteger: BigInt?, biginteger_14_: BigInt?) {
     private val aClass248_4212: Class248?
-    private val aBigInteger4213: BigInteger?
+    private val aBigInteger4213: BigInt?
     private var aClass348_Sub49_4215: Class348_Sub49? = null
     private var aClass314_Sub1Array4216: Array<Class314_Sub1?>? = null
     private var aClass112_4217: Class112? = null
-    private val aBigInteger4218: BigInteger?
+    private val aBigInteger4218: BigInt?
     private var aClass348_Sub42_Sub16_Sub1_4220: Class348_Sub42_Sub16_Sub1? = null
     fun method2670(i: Int) {
         anInt4222++
@@ -55,7 +54,7 @@ class Class340 internal constructor(class248: Class248?, class112: Class112?, bi
         class348_sub49.method3389(2147483647, 0, `is`.size, `is`)
         val is_6_: ByteArray?
         if (aBigInteger4213 != null && aBigInteger4218 != null) {
-            val biginteger = BigInteger(`is`)
+            val biginteger = BigInt(`is`)
             val biginteger_7_ = biginteger.modPow(aBigInteger4213, aBigInteger4218)
             is_6_ = biginteger_7_.toByteArray()
         } else {
